Move DateControl to a shared component
